requirements.txt
openpyxl==3.1.2
pandas==1.5.1
python-dotenv==1.0.0
requests==2.28.1